investor onboard page
