+++ tinlake-ui/components/Onboarding/PlainCheckIcon.tsx
@@ -1,0 +1,9 @@
+import * as React from 'react'
+
+export const PlainCheckIcon: React.FC = () => {
+  return (
+    <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <path d="M20 6L9 17L4 12" stroke="black" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" />
+    </svg>
+  )
+}
